% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{assign_vars}
\alias{assign_vars}
\title{Assign result to variables}
\usage{
assign_vars(result, vars)
}
\arguments{
\item{result}{A named list with results}

\item{vars}{A list of parameter \code{\link{Variable}} objects}
}
\description{
Assigns results back to decision variable objects.
}
\keyword{internal}
