% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraint.R
\docType{data}
\name{Constraint}
\alias{Constraint}
\title{Constraint Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Constraint
}
\description{
Constraint Class
}
\keyword{datasets}
