% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{.assign_vars}
\alias{.assign_vars}
\title{Assign result to variables}
\usage{
.assign_vars(result)
}
\arguments{
\item{result}{A named list with results}
}
\description{
Assigns results back to decision variable objects.
}
