% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model}
\alias{Model}
\title{Model Class}
\description{
Model is a class to describe a Minizinc model.
It contains information that is needed to define a Minizinc model, such as
a list of variables, constraints and model objective (e.g. satisfy).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parameter}}{A list of parameter \code{\link{Variable}} objects.}

\item{\code{decision}}{A list of decision \code{\link{Variable}} objects.}

\item{\code{constraints}}{A list of \code{\link{Constraint}} objects.}

\item{\code{objective}}{Model objective, which can be either of satisfy, maximize or minimize.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Model$new()}}
\item \href{#method-clone}{\code{Model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of Model class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$new(
  parameter = NULL,
  decision = NULL,
  constraints = NULL,
  objective = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
