% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{print_vars}
\alias{print_vars}
\title{Print variable values}
\usage{
print_vars(vars)
}
\arguments{
\item{vars}{A list of \code{\link{Variable}} objects}
}
\description{
Given a list of variables, this function prints their values.
}
