% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\name{Solver}
\alias{Solver}
\title{Solver Class}
\description{
Solver is a class to describe a Minizinc solver.
It contains information that is needed to define a Minizinc solver, such as
a name of the solver and its configuration.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character name of solver. Default is "gecode".}

\item{\code{config}}{A list of configuration passed to Minizinc solver.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Solver$new()}}
\item \href{#method-clone}{\code{Solver$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of Solver class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solver$new(name = NULL, config = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Solver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
