% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable.R
\docType{data}
\name{Variable}
\alias{Variable}
\title{Variable Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Variable
}
\description{
Variable is a class to describe both parameter and decision variables.
It contains information that is needed to define a Minizinc variable, such as
type (int, bool, float, etc),
kind (decision, parameter),
value
domain
}
\keyword{datasets}
