% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{write_cmd}
\alias{write_cmd}
\title{Write system command}
\usage{
write_cmd(path, solver)
}
\arguments{
\item{path}{A string with Minizinc path}

\item{solver}{\code{\link{Solver}} object}
}
\description{
This function builds a system command to execute.
}
\keyword{internal}
