% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{minizinc-package}
\alias{minizinc}
\alias{minizinc-package}
\title{minizinc: Object Oriented Interface to Minizinc in R}
\description{

}
\author{
\strong{Maintainer}: Damir Pulatov \email{dpulatov@uwyo.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
