% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{minizinc}
\alias{minizinc}
\alias{.globals}
\title{R Interface to Minizinc}
\format{
An object of class \code{environment} of length 5.
}
\usage{
.globals
}
\description{
R interface to Minizinc language.
}
\keyword{datasets}
