% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve}
\alias{solve}
\title{Solve models}
\usage{
solve(model, solver)
}
\arguments{
\item{model}{\code{\link{Model}}
Object of class Model.}

\item{solver}{Object of class Solver.}
}
\value{
list with results
}
\description{
This function is used to solve a given model. The function returns
solution in list format.
}
\examples{
# Create and solve a Minizinc model
var1 = Variable$new(type = "int", kind = "decision", domain = c(0, 10))
var1 = Variable$new(type = "int", kind = "decision", domain = c(0, 5))

constr = Constraint$new(constraint = "<", variables = c(var1, var2))

model = Model$new(decision = c(var1, var2), constraints = c(constr),
                  objective = "satisfy")
solver = Solver$new(name = "gecode")

res = solve(model, solver)
print(res)
}
