% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve}
\alias{solve}
\title{Solve models}
\usage{
solve(model, solver)
}
\arguments{
\item{model}{\code{\link{Model}}
Object of class Model.}

\item{solver}{Object of class Solver.}
}
\value{
list with results
}
\description{
This function is used to solve a given model. The function returns
solution in list format.
}
