% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve}
\alias{solve}
\title{Solve models}
\usage{
solve(model, solver = "gecode")
}
\arguments{
\item{model}{[\code{Model}]\cr
Model that contains variables, constraints and objective.}

\item{Solver}{object that specifies both solver and its configuration. Default is gecode.}
}
\value{
data.frame object with results
}
\description{
This function is used to solve a given model.
}
