% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variable.R
\name{Variable}
\alias{Variable}
\title{Variable Class}
\description{
Variable is a class to describe both parameter and decision variables.
It contains information that is needed to define a Minizinc variable, such as
type (int, bool, float, etc), kind (decision, parameter), value, domain.
}
\section{Construction}{


```
v = Variable$new(type, kind, value, domain)
```

* `type` :: `character(1)`\cr
Type of the variable.

* `kind` :: `character(1)`\cr
Variable kind.

* `kind` :: `character(1)`\cr
Value of the variable.

* `domain` :: `list(1)`\cr
Variable domain.

Creates a new instance of this [R6][R6::R6Class] class.
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{Type of variable (int, float, bool, etc.).}

\item{\code{kind}}{Kind of variable (parameter, decision).}

\item{\code{value}}{Value assigned to this variable. Value is only meaningful for
parameter variables.}

\item{\code{domain}}{Domain of valid values for a decision variable. Domain is only
meaningful for decision variables.}

\item{\code{name}}{Name of variable as it will appear in Minizinc.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Variable$new()}}
\item \href{#method-clone}{\code{Variable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$new(type, kind, value = NULL, domain = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Variable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
